requirements.txt
annotated-types==0.7.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
Flask==3.1.2
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pydantic==2.12.3
pydantic_core==2.41.4
PyYAML==6.0.3
requests==2.32.5
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
zipp==3.23.0
psutil==7.1.3